Implement: PursuitBehavior.cs, Update PursuitEnemy
